+++ CHANGELOG.md
@@ -24,7 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 - `Movie#play()` now waits until all resources are loaded to start playing.
-- Array methods like `unshift` for `etro.layer.Visual#effects` work properly.
+- Array methods like `unshift` for `etro.layer.Visual#effects` and `etro.effect.Stack#effects` work properly.
 - Too many WebGL contexts error when using more than 16 effects.
 - `AudioSource#source` is now typed as an `HTMLAudioElement` ([#168](https://github.com/etro-js/etro/pull/168)).